- #4 RoutingModel and RoutingMessage separation from OTPRouter - Refactor imports to new hierarchy and organize them
