+++ src/main/scala/beam/agentsim/package.scala
@@ -1,10 +1,8 @@
 package beam
 
 import beam.agentsim.agents.PersonAgent
-import beam.agentsim.core.Modes.BeamMode
 import org.matsim.api.core.v01.Id
 import org.matsim.api.core.v01.population.Person
-import org.opentripplanner.routing.core.TraverseMode
 
 /**
   * Created by sfeygin on 1/27/17.
