+++ src/main/scala/beam/agentsim/playground/rashid/SimpleActor.scala
@@ -1,8 +1,6 @@
 package beam.agentsim.playground.rashid
 
-import akka.actor.Actor
-import akka.actor.ActorSystem
-import akka.actor.Props
+import akka.actor.{Actor, ActorSystem, Props}
 
 object SimpleActor extends App {
