+++ src/main/scala/beam/agentsim/playground/rashid/physicalSim/jdeqsim/oneZone/ActorVersionJDEQSim.scala
@@ -1,17 +1,8 @@
 package beam.agentsim.playground.rashid.physicalSim.jdeqsim.oneZone
 
-import akka.actor.Actor
-import akka.actor.ActorSystem
-import akka.actor.Props
-
+import akka.actor.{Actor, ActorSystem, Props}
 import org.matsim.core.config.ConfigUtils
 import org.matsim.core.scenario.ScenarioUtils
-import org.matsim.core.events.EventsUtils
-import beam.playground.jdeqsim.CountEnterLinkEvents
-import org.matsim.core.mobsim.jdeqsim.JDEQSimConfigGroup
-import org.matsim.core.mobsim.jdeqsim.JDEQSimulation
-import org.matsim.core.events.EventsManagerImpl
-import beam.playground.jdeqsim.akka.AkkaEventHandlerAdapter
 
 object ActorVersionJDEQSim {
