+++ src/main/scala/beam/agentsim/routing/BeamRouterModuleProvider.scala
@@ -1,9 +1,7 @@
 package beam.agentsim.routing
 
-import beam.agentsim.routing.opentripplanner.OpenTripPlannerRouter
 import beam.agentsim.sim.AgentsimServices
 import com.google.inject.{Inject, Provider}
-import glokka.Registry
 import org.matsim.core.router.{RoutingModule, TripRouter}
 
 /**
