+++ src/main/scala/beam/agentsim/routing/RoutingMessages.scala
@@ -1,6 +1,7 @@
 package beam.agentsim.routing
 
 import beam.agentsim.agents.PersonAgent
+import beam.agentsim.routing.RoutingModel.BeamTrip
 import org.matsim.api.core.v01.Id
 import org.matsim.api.core.v01.population.Activity
 import org.matsim.core.router.ActivityWrapperFacility
@@ -19,4 +20,6 @@ object RoutingMessages {
       new RoutingRequest(new ActivityWrapperFacility(fromActivity), new ActivityWrapperFacility(toActivity), departureTime, personId)
     }
   }
+
+  case class RoutingResponse(itinerary: Vector[BeamTrip])
 }
