+++ src/main/scala/beam/agentsim/routing/opentripplanner/OpenTripPlannerRouter.scala
@@ -9,10 +9,9 @@ import akka.actor.Props
 import beam.agentsim.config.BeamConfig
 import beam.agentsim.core.Modes.BeamMode
 import beam.agentsim.core.Modes.BeamMode._
-import beam.agentsim.events.SpaceTime
 import beam.agentsim.routing.BeamRouter
 import beam.agentsim.routing.RoutingMessages._
-import beam.agentsim.routing.opentripplanner.OpenTripPlannerRouter._
+import beam.agentsim.routing.RoutingModel.{BeamGraphPath, BeamLeg, BeamTrip, EdgeModeTime}
 import beam.agentsim.sim.AgentsimServices
 import beam.agentsim.utils.GeoUtils
 import beam.agentsim.utils.GeoUtils._
@@ -293,54 +292,4 @@ class OpenTripPlannerRouter @Inject() (agentsimServices: AgentsimServices, beamC
 
 object OpenTripPlannerRouter {
   def props(agentsimServices: AgentsimServices) = Props(classOf[OpenTripPlannerRouter], agentsimServices)
-
-  case class RoutingResponse(itinerary: Vector[BeamTrip])
-
-  case class BeamTrip(legs: Vector[BeamLeg], choiceUtility: Double = 0.0) {
-    lazy val tripClassifier: BeamMode = if (legs map (_.mode) contains CAR) {
-      CAR
-    } else {
-      TRANSIT
-    }
-    val totalTravelTime: Long = legs.map(_.travelTime).sum
-
-  }
-
-  object BeamTrip {
-    val noneTrip: BeamTrip = BeamTrip(Vector[BeamLeg]())
-
-  }
-
-  case class BeamLeg(startTime: Long, mode: BeamMode, travelTime: Long, graphPath: BeamGraphPath)
-
-  object BeamLeg {
-
-    def dummyWalk(startTime: Long): BeamLeg = new BeamLeg(startTime, WALK, 0, BeamGraphPath.empty)
-
-  }
-
-  case class BeamGraphPath(linkIds: Vector[String],
-                           latLons: Vector[Coord],
-                           entryTimes: Vector[Long]) {
-
-    lazy val trajectory: Vector[SpaceTime] = {
-      latLons zip entryTimes map {
-        SpaceTime(_)
-      }
-    }
-  }
-
-  object BeamGraphPath {
-    val emptyTimes: Vector[Long] = Vector[Long]()
-    val errorPoints: Vector[Coord] = Vector[Coord](new Coord(0.0, 0.0))
-    val errorTime: Vector[Long] = Vector[Long](-1L)
-
-    val empty: BeamGraphPath = new BeamGraphPath(Vector[String](), errorPoints, emptyTimes)
-
-
-  }
-
-  case class EdgeModeTime(fromVertexLabel: String, mode: BeamMode, time: Long, fromCoord: Coord, toCoord: Coord)
-
-
 }
